cl-deploy: Update source block device logics
part_dump logics update)
1) Remove the last partition and the last-lba from the source dump;
it allows dumping to a low-sized media.
In this case it is up to the operator to handle: a not enough space failure.
2) Append the source last partition to the dst with the open-end block:
to the end of the available disk's space.

source_block_device logics update)
Remove the external last partition expand logics. The part_dump() does it.

Signed-off-by: Valentin Raevsky <valentin@compulab.co.il>